requirements.txt
flask 
line-bot-sdk
configparser 
openai
azure-cognitiveservices-speech
librosa
